refactor: return useStreamEventListener from EventStreamContext
