+++ src/modules/event-stream/EventStreamContext.tsx
@@ -1,28 +1,22 @@
-import {createContext, useContext} from 'react';
-import {
-  StreamEventListener,
-  useSetupEventStream,
-} from './use-setup-event-stream';
-import {EventKind, StreamEventLog} from './types';
+import {createContext, useContext, useEffect, useId, useRef} from 'react';
+import {useSetupEventStream} from './use-setup-event-stream';
+import {EventKind, StreamEventLog, StreamEventOfKind} from './types';
 
 interface EventStreamContextValue {
   eventLog: StreamEventLog;
   /**
-   * NOTE: Usually should not be called directly, but rather through the
-   * `useEventStreamListener` hook.
+   * Listens for events of a specific kind from the event stream, and calls the
+   * provided callback when events of that kind are received.
    */
-  subscribe: <K extends EventKind>(listener: StreamEventListener<K>) => void;
-  /**
-   * NOTE: Usually should not be called directly, but rather through the
-   * `useEventStreamListener` hook.
-   */
-  unsubscribe: (id: string) => void;
+  useStreamEventListener: <K extends EventKind>(
+    eventKind: K,
+    callback: (streamEvent: StreamEventOfKind<K>) => void,
+  ) => void;
 }
 
 export const EventStreamContext = createContext<EventStreamContextValue>({
   eventLog: [],
-  subscribe: () => {},
-  unsubscribe: () => {},
+  useStreamEventListener: () => {},
 });
 
 export function useEventStreamContext() {
@@ -42,12 +36,40 @@ type Props = {
 export const EventStreamContextProvider = ({children}: Props) => {
   const {eventLog, subscribe, unsubscribe} = useSetupEventStream();
 
+  function useStreamEventListener<K extends EventKind>(
+    eventKind: K,
+    callback: (streamEvent: StreamEventOfKind<K>) => void,
+  ) {
+    const id = useId();
+
+    // Keep the callback in a ref to avoid resubscribes on rerenders.
+    const callbackRef = useRef(callback);
+    useEffect(() => {
+      callbackRef.current = callback;
+    }, [callback]);
+
+    useEffect(() => {
+      // The event kind isn't unique when several listeners listen to the same
+      // event kind, so we combine it with an ID that is unique to this hook
+      // instance.
+      const listenerId = `${eventKind}-${id}`;
+
+      subscribe({
+        id: listenerId,
+        eventKind,
+        callback: (streamEvent) => callbackRef.current(streamEvent),
+      });
+      return () => {
+        unsubscribe(listenerId);
+      };
+    }, [eventKind, id]);
+  }
+
   return (
     <EventStreamContext.Provider
       value={{
         eventLog,
-        subscribe,
-        unsubscribe,
+        useStreamEventListener,
       }}
     >
       {children}
